refactor: Participation 도메인 관련 폴더 구조 리팩토링
