+++ frontend/src/components/participation/ParticipationCodeCopier/ParticipationCodeCopier.tsx
