+++ frontend/src/hooks/participation/useRegisterMember.ts
