+++ frontend/src/pages/CreateStudy.tsx
@@ -1,12 +1,12 @@
 import CreateStudyForm from '@Components/create/CreateStudyForm/CreateStudyForm';
 
-import StudyPreparationLayout from './layout/StudyPreparationLayout';
+import StudyParticipationLayout from './layout/StudyParticipationLayout';
 
 const CreateStudy = () => {
   return (
-    <StudyPreparationLayout headerText="스터디 개설하기">
+    <StudyParticipationLayout headerText="스터디 개설하기">
       <CreateStudyForm />
-    </StudyPreparationLayout>
+    </StudyParticipationLayout>
   );
 };
